oci: split pulling code to new file, skopeo.rs
This code pulls images via containers-image-proxy-rs (which is a
wrapper around skopeo).  It's a bit too large to have it in the
top-level lib.rs, and we're about to add an alternate implementation, so
split it out.

We keep a pull() function in place for compatibility reasons to avoid
needing to update all callers: we will start using this function to
dispatch to the correct implementation depending on the reference
format.
